rename file so it's actually picked up by jest and fix the test
